Visual design and Layout improvements   * add wip placeholder light theme   * iterate dark theme   * create pair-card component
