+++ packages/client/src/components/card/pair-card.scss
@@ -1,0 +1,10 @@
+
+.pair-card {
+        margin-right: 1rem;
+        margin-top: 1rem;
+        padding: 1rem 1.4rem;
+        width: 15rem;
+        border: 1px solid var(--borderMain);
+        box-shadow: 0 0 5px 1px var(--bgDeep);
+        background: var(--bgDeep);
+    }
